+++ .gitignore
@@ -1,3 +1,4 @@
-+.gitignore
- +.idea/
- +/target/ 
+.classpath
+.project
+.settings/
+target/
