+++ src/main/java/me/F_o_F_1092/TimeVote/TimerType.java
